Moved CityReligion to its own file
